+++ tests/indicators/_common/Series/Test.Transforms.cs
@@ -1,8 +1,9 @@
 using System.Collections.ObjectModel;
+using Internal.Tests;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Skender.Stock.Indicators;
 
-namespace Internal.Tests;
+namespace Tests.Indicators;
 
 [TestClass]
 public class TransformTests : TestBase
